requirements.txt
tensorflow-cpu==
transformers==4.11.3
pandas==1.5.3
matplotlib==3.4.3